Deploy Calico before Multus in nic-operator CI
Deploying Calico CNI after the Multus CNI causes the Calico
Controller to always fail, though this does not affect workloads
with secondary networks, it affects the coredns pods causing them
to fail to start. This patch changes the order of deploying the CNIs
in the nic-operator so that Calico will be deployed first before
the Multus to run the coredns, necessary for NFD deployment.